# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DDC-API API Document",
        "version": "1.0.0"
    },
    "paths": {
        "/api/getDisease/{disease_id}": {
            "get": {
                "tags": [
                    "Disease"
                ],
                "summary": "Count Disease",
                "operationId": "Disease Count",
                "parameters": [
                    {
                        "name": "disease_id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "type": "bigint",
                                            "example": "9999"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Retrieved Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/getDiseaseGender/{disease_id}": {
            "get": {
                "tags": [
                    "Disease"
                ],
                "summary": "Count Gender",
                "operationId": "Disease Gender",
                "parameters": [
                    {
                        "name": "disease_id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "male": {
                                            "type": "bigint",
                                            "example": "9999"
                                        },
                                        "female": {
                                            "type": "bigint",
                                            "example": "9999"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Retrieved Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/getDiseaseProvince/{disease_id}": {
            "get": {
                "tags": [
                    "Disease"
                ],
                "summary": "Count Province",
                "operationId": "Disease Province",
                "parameters": [
                    {
                        "name": "disease_id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "PROVINCE": {
                                            "type": "bigint",
                                            "example": "10"
                                        },
                                        "count_province": {
                                            "type": "bigint",
                                            "example": "9999"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Retrieved Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/getDiseaseAge/{disease_id}": {
            "get": {
                "tags": [
                    "Disease"
                ],
                "summary": "Count Age",
                "operationId": "Disease Age",
                "parameters": [
                    {
                        "name": "disease_id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "newborn": {
                                            "type": "bigint",
                                            "example": "9999"
                                        },
                                        "children": {
                                            "type": "bigint",
                                            "example": "9999"
                                        },
                                        "working": {
                                            "type": "bigint",
                                            "example": "9999"
                                        },
                                        "senile": {
                                            "type": "bigint",
                                            "example": "9999"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Retrieved Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}